requirements/base.txt
Django==1.8.2
django-braces==1.8.0
django-flat-theme==0.9.5
Pillow==2.8.2
psycopg2==2.6.1
six==1.9.0
tablib==0.10.0
Unipath==1.1
